+++ app/components/generic/async-button.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/generic/async-button';